changed the cache redis expiration value type to int64 for seconds

+++ cache/common.go
@@ -5,17 +5,17 @@ import "encoding/gob"
 
 // Cacher defines the interface for all typs of cacher. e.g. redis, memcached and etc.
 type Cacher interface {
-	Set(key string, value interface{}, expiration ...int) error
+	Set(key string, value interface{}, expiration ...int64) error
 	Get(key string) (interface{}, error)
 	Del(key string)
 	Scan(cursor int, count int, pattern string) (nextCursor int, keys []string, err error)
 	Expire(key string, expiration int) error
 	TTL(key string) (int, error)
-	SetGob(key string, value interface{}, expiration ...int)
+	SetGob(key string, value interface{}, expiration ...int64)
 	GetGob(key string) (interface{}, error)
-	SetJSON(key string, value interface{}, expiration ...int)
+	SetJSON(key string, value interface{}, expiration ...int64)
 	GetJSON(key string) (jsonBytes []byte, err error)
-	HSet(hash, key string, value interface{}, expiration ...int) error
+	HSet(hash, key string, value interface{}, expiration ...int64) error
 	HGet(hash, key string) (interface{}, error)
 	HINCRBY(hash, key string, value interface{}) error
 }
